Finished with arguments (minimum spec)
Enable Assertions implemented correctly (has a test example in a
previous commit)

Arguments are passed in through the command line (as well as enable
assertions) and will show up on Deeva

Arguments can be changed on Deeva and then that affects the next
program state.

Arguments are saved so if you open up a new tab it will keep those
arguments.

Also some code tidy.